+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project are documented here. The format
 is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
-## [Unreleased]
+## [0.18.2] - 2026-08-22
 
 ### Fixed